Add files via upload
## Human Robot Interaction

# NAO robot execise coach

A Human Robot Interaction project where a NAO robot (https://www.aldebaran.com/en/nao) is programmed to interact with the user and act as a coach 
for some simple exercises such as squats, arm raises, arm circles etc. The aim of the project is to investigate how useful a social robot could
be for patients doing physical movements.